+++ src/app/baniSupport/matra.js
@@ -1,0 +1,5 @@
+const matra = {
+  japji: {},
+};
+
+export default matra;