+++ src/app/config.js
@@ -1,8 +1,10 @@
 import lineOrder from './baniSupport/lineOrder';
+import matra from './baniSupport/matra';
 import sectionOrder from './baniSupport/sectionOrder';
 
 export const supportedBanis = {
   lineOrder,
+  matra,
   sectionOrder,
 };